# Add IteratorCursor and getMessageCursor to iterable player architecture