--- a/packages/studio-base/src/players/IterablePlayer/BlockLoader.ts
+++ b/packages/studio-base/src/players/IterablePlayer/BlockLoader.ts
@@ -60,6 +60,7 @@ export class BlockLoader {
   private problemManager: PlayerProblemManager;
   private stopped: boolean = false;
   private activeChangeCondvar: Condvar = new Condvar();
+  private abortController: AbortController;
 
   public constructor(args: BlockLoaderArgs) {
     this.source = args.source;
@@ -67,6 +68,7 @@ export class BlockLoader {
     this.end = args.end;
     this.maxCacheSize = args.cacheSizeBytes;
     this.problemManager = args.problemManager;
+    this.abortController = new AbortController();
 
     const totalNs = Number(toNanoSec(subtractTimes(this.end, this.start))) + 1; // +1 since times are inclusive.
     if (totalNs > Number.MAX_SAFE_INTEGER * 0.9) {
@@ -92,6 +94,7 @@ export class BlockLoader {
       return;
     }
 
+    this.abortController.abort();
     this.activeBlockId = beginBlockId;
     this.activeChangeCondvar.notifyAll();
   }
@@ -101,6 +104,7 @@ export class BlockLoader {
       return;
     }
 
+    this.abortController.abort();
     this.topics = topics;
     this.activeChangeCondvar.notifyAll();
 
@@ -129,6 +133,8 @@ export class BlockLoader {
 
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     while (!this.stopped) {
+      this.abortController = new AbortController();
+
       const activeBlockId = this.activeBlockId;
       const topics = this.topics;
 
@@ -157,23 +163,15 @@ export class BlockLoader {
       return;
     }
 
-    // We prioritize loading from the active block id to the end, then we load from the start to active block id
-    const segments: [number, number][] = [
-      [this.activeBlockId, this.blocks.length],
-      [0, this.activeBlockId],
-    ];
+    if (this.blocks.length === 0) {
+      return;
+    }
 
-    log.debug("loading segments", segments);
+    // Load the active block to end first, then go back and load first block to active block
+    await this.loadBlockRange(this.activeBlockId, this.blocks.length, args.progress);
 
-    for (const segment of segments) {
-      const [beginBlockId, lastBlockId] = segment;
-
-      // Note: lastBlockId is actually 1-past-last block so we can skip this loop if the begin and last are the same
-      if (beginBlockId === lastBlockId) {
-        continue;
-      }
-
-      await this.loadBlockRange(beginBlockId, lastBlockId, args.progress);
+    if (this.activeBlockId > 0) {
+      await this.loadBlockRange(0, this.activeBlockId, args.progress);
     }
   }
 
@@ -239,17 +237,20 @@ export class BlockLoader {
       // If the source provides a message cursor we use its message cursor, otherwise we make one
       // using the source's message iterator.
       const cursor =
-        this.source.getMessageCursor?.(iteratorArgs) ??
-        new IteratorCursor(this.source.messageIterator(iteratorArgs));
+        this.source.getMessageCursor?.({ ...iteratorArgs, abort: this.abortController.signal }) ??
+        new IteratorCursor(this.source.messageIterator(iteratorArgs), this.abortController.signal);
 
       for (let currentBlockId = blockId; currentBlockId <= endBlockId; ++currentBlockId) {
         const untilTime = clampTime(this.blockIdToEndTime(currentBlockId), this.start, this.end);
 
         const results = await cursor.readUntil(untilTime);
+        // No results means cursor aborted or eof
+        if (!results) {
+          return;
+        }
 
         const messagesByTopic: Record<string, MessageEvent<unknown>[]> = {};
-
-        if (!results || results.length === 0) {
+        if (results.length === 0) {
           // Set all topics to empty arrays since this block has no messages on these topics
           for (const topic of topicsToFetch) {
             messagesByTopic[topic] = [];
@@ -323,16 +324,6 @@ export class BlockLoader {
           },
           sizeInBytes: (existingBlock?.sizeInBytes ?? 0) + sizeInBytes,
         };
-
-        // fixme - readuntil should have an abort controller since previously we could abort after a single message read
-        // if we abourt part way we need to avoid updating the block!
-        // now we can only abort at the readUntil level
-        // When the topics change we bail and will re-load from the outer loading loop
-        // We do this after having the results
-        if (topics !== this.topics) {
-          log.debug("topics changed, aborting load instance");
-          return;
-        }
 
         // fixme
         // Emitting after every block results in fighting with the plot panel