… packages/comlink-transfer-handlers/src/iterableTransferHandler.ts
… packages/studio-base/src/players/IterablePlayer/IteratorCursor.ts
… packages/studio-base/src/players/IterablePlayer/WorkerIterableSource.ts
… packages/studio-base/src/players/IterablePlayer/WorkerIterableSourceWorker.worker.ts