@defunctzombie 2022-10-20
> But do we also plan to eventually use readUntil for regular playback when we are loading a frame of messages?

Yes - I wanted to keep this change focused on block loader to minimize potential disruption and the diff if there is some issue. I think it would help regular playback just a pinch.

> Looks like you still have some fixmes to resolve.

Indeed. My plan was to resolve those after the first round of comments.